separate `crate::utils` and `crate::view::utils`
- cleanup imports
