+++ hackernews_tui/src/client/parser.rs
@@ -1,4 +1,5 @@
 use crate::prelude::*;
+use crate::utils;
 use lazy_static::lazy_static;
 use rayon::prelude::*;
 use regex::Regex;
