+++ hackernews_tui/src/client/query.rs
@@ -1,4 +1,4 @@
-use crate::prelude::*;
+use crate::utils;
 use serde::Deserialize;
 
 #[derive(Debug, Clone, Copy, Default, Deserialize)]
