+++ hackernews_tui/src/view/search_view.rs
@@ -1,4 +1,4 @@
-use super::{help_view::*, story_view, text_view::EditableTextView};
+use super::{help_view::*, story_view, text_view::EditableTextView, utils};
 use crate::prelude::*;
 
 #[derive(Copy, Clone, PartialEq, Eq)]
